requirements.txt
Flask==2.1.0
Flask-MySQL==1.5.2
